pyproject.toml
[tool.poetry]
name = "deep-domain-detection"
version = "1.1.2"
description = "Classification pipeline"
authors = ["Jan Polisensky <Polisensky.john@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
scikit-learn = "^1.5.0"
sklearn-pandas = "^2.2.0"
tensorflow-io-gcs-filesystem = {version = ">=0.36.0", python = ">=3.11,<3.12"}
tensorflow-cpu = "~=2.15.0"
keras = "^2.15.0"
xgboost = "^2.0.3"
lightgbm = "^4.3.0"
pandas = ">=2.0.0"
numpy = ">=1.24.3"
python-math = "^0.0.1"
joblib = "^1.3.2"
dill = "^0.3.8"
shap = "~=0.45.1"
torch = "^2.1.2"

[tool.poetry.group.arrow]
optional = true

[tool.poetry.group.arrow.dependencies]
pyarrow = "^15.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
notebook = ">=6.5.4"
pymongoarrow = ">=0.7.0"
pyarrow = "^15.0.0"
scipy = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"